fix bug 2

+++ contracts/GLXRouter.sol
@@ -69,7 +69,7 @@ contract GLXRouter is Ownable {
         //同样的押注资金：在同一个epoch中，越早押注铸币越多。在不同epoch中，同一个押注次序，早期的epoch会收获更多的代币
         uint256 totalMintAmount = amount.div( (uint256(1).add(curUserCount.div(uint256(10))) )).mul(uint256(100)).div(discount);
 
-        uint256 liquidMintAmount = totalMintAmount.mul(uint256(LIQUID_MINT_RATE)).div(uint256(100));
+        uint256 liquidMintAmount = totalMintAmount.mul(LIQUID_MINT_RATE).div(uint256(100));
         uint256 userMintAmount = totalMintAmount.sub(liquidMintAmount);
 
         //为股东铸币
